// pslpwcl/src/app/modules/pslpapi/model/appuntamentiRidotta.ts
/*
 * ========================LICENSE_START=================================
 * Copyright (C) 2025 Regione Piemonte
 * SPDX-FileCopyrightText: Copyright 2025 | Regione Piemonte
 * SPDX-License-Identifier: EUPL-1.2
 * =========================LICENSE_END==================================
 */
/**
 * 
 * 
 *
 * 
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export interface AppuntamentiRidotta { 
    altroTelefonoSedeIncontro?: string;
    capSedeIncontro?: string;
    cellulareAnagrafica?: string;
    codFiscaleAnagrafica?: string;
    codTipoAgendaIncServizi?: string;
    cognomeAnagrafica?: string;
    cognomeOperatore?: string;
    dataIncontro?: Date;
    dataLimiteSpostamento?: Date;
    desComeRaggiungereSedeIncontro?: string;
    desComuneSedeIncontro?: string;
    desCpiOperatore?: string;
    desCpiPromotoreAnagrafica?: string;
    desEstesaSilwebTIncServizi?: string;
    desFontedatoIncontro?: string;
    desProvinciaSedeIncontro?: string;
    desSedeCpiIncontro?: string;
    desSedeIncontro?: string;
    desSilDTipoSedeIncontro?: string;
    desSilTInContattoErogaz?: string;
    desSilTInMotivazione?: string;
    desSilTStatoIncontro?: string;
    desSilwebTIncServCateg?: string;
    desSilwebTIncServizi?: string;
    desSocietaOperatore?: string;
    desTestoAnnulla?: string;
    desTestoSposta?: string;
    desToponimoSedeIncontro?: string;
    dsNoteIncontro?: string;
    durataEffettivaIncontroMinuti?: number;
    durataPrevistaIncontroMinuti?: number;
    flgPrenotaCittadinoServ?: string;
    flgSpostaServ?: string;
    idComuneSedeIncontro?: string;
    idCpiOperatore?: number;
    idCpiPromotoreAnagrafica?: number;
    idEnteOperatore?: number;
    idFontedatoIncontro?: number;
    idIncontro?: number;
    idIncontroPrec?: number;
    idMsgDisdettaOperSilos?: string;
    idMsgNonPresentatoSilos?: string;
    idMsgPrenotazSilos?: string;
    idMsgPromemoriaSilos?: string;
    idOperatore?: string;
    idProvinciaSedeIncontro?: string;
    idSedeCpiIncontro?: number;
    idSedeIncontro?: number;
    idSilDTipoSedeIncontro?: number;
    idSilLavAnagrafica?: number;
    idSilTInContattoErogaz?: number;
    idSilTInMotivazione?: number;
    idSilTStatoIncontro?: string;
    idSilwebTIncServCateg?: number;
    idSilwebTIncServizi?: number;
    idSocietaOperatore?: number;
    idToponimoSedeIncontro?: string;
    indirizzoSedeIncontro?: string;
    mailAnagrafica?: string;
    mailSedeIncontro?: string;
    nomeAnagrafica?: string;
    nomeOperatore?: string;
    numCivicoSedeIncontro?: string;
    numGiorniSposta?: number;
    oraIncontro?: string;
    subCodiceSedeIncontro?: number;
    telefonoSedeIncontro?: string;
    totFasce?: number;
    totFascePresenza?: number;
    totFasceRemoto?: number;
}