// pslpwcl/src/app/modules/pslpapi/model/dettaglioIncontro.ts
/*
 * ========================LICENSE_START=================================
 * Copyright (C) 2025 Regione Piemonte
 * SPDX-FileCopyrightText: Copyright 2025 | Regione Piemonte
 * SPDX-License-Identifier: EUPL-1.2
 * =========================LICENSE_END==================================
 */
/**
 * 
 * 
 *
 * 
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { AppuntamentiRidotta } from './appuntamentiRidotta';
import { IncServizi } from './incServizi';
import { SilwebTStoServIncontri } from './silwebTStoServIncontri';

export interface DettaglioIncontro { 
    incontro?: AppuntamentiRidotta;
    listaSpostamentiPrec?: Array<AppuntamentiRidotta>;
    listaStorico?: Array<SilwebTStoServIncontri>;
    servizio?: IncServizi;
}